Codeclimate fix - moved file logic into a new FileUtils
