+++ src/main/java/ca/bc/gov/ag/efax/mail/service/EmailService.java
@@ -2,6 +2,7 @@ package ca.bc.gov.ag.efax.mail.service;
 
 import java.util.List;
 
+import ca.bc.gov.ag.efax.mail.model.MailMessage;
 import microsoft.exchange.webservices.data.core.service.item.EmailMessage;
 
 public interface EmailService {
@@ -13,5 +14,12 @@ public interface EmailService {
      * @throws Exception 
      */
     List<EmailMessage> getEfaxInboxEmails() throws Exception;
+    
+    /**
+     * Sends an email (including any attachments) using Microsoft Exchange.
+     * 
+     * @param mailMessage
+     */
+    public void sendMessage(final MailMessage mailMessage);
     
 }
